+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2025-01-05
+- #2141 adapts the `sov-value-setter` module to benchmark generation. This is not a breaking change for customers of the SDK.
+
 ## 2025-01-04
 - #2137 introduces the FromVmAddress trait bounds to demo-rollup (which includes an EVM module), allowing EVM transactions to be submitted natively from EVM addresses—without requiring the registration of EVM addresses as credentials.